Prunning Test cases for ...
